requirements.txt
Django==5.1.7
rq==2.1
djangorestframework==3.15.2
django-environ==0.11.2
django-filter
feedparser==6.0.11
requests==2.32.3
